SOA: add tools dir
SOA: sync dump_obj

SOA: update tools

SOA: add compile.bat

SOA: remove config.in for tools

SOA: use global nc2fc.py

+++ src/custom/soa/rewrite_soa_vgame_en/soa_seg008.h
@@ -3,7 +3,7 @@ namespace SOA_VGAME_EN {
 #endif
 
 unsigned short random_interval(const unsigned short, const unsigned short);
-unsigned short random_soa(const unsigned short);
+signed int random_soa(const signed int);
 unsigned short is_in_word_array(const signed short, signed short*);
 unsigned short is_in_byte_array(const signed char, signed char*);
